Put code in functions, use rel paths, correct distance calculation
Distance calculation incorrectly used previous distance when no movement was recorded.

Wrap in functions and write csv to new file, rather than renaming a copy of the csv before manipulating. This also makes it easy to use the timestamp recorded in the csv file, which is corrected for the timezone.

Use `{pyprojroot}` package for relative paths